pyproject.toml
[tool.poetry]
name = "rebus-project"
version = "0.1.0"
description = ""
authors = ["Denis Gonzalez <deniseduardoisidrogonzalez@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.4"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
